LayoutTests/fast/shadow-dom/mouseenter-mouseleave-inside-shadow-tree-expected.txt

This tests mouseenter and mouseleave events inside a shadow tree.
To manually test, move the mouse cursor vertically across "hello, world" below.
WebKit must generate mouseenter and mouseleve events exactly once on host (as well as hostParent, continer, and target).


==Entering target==
mouseenter on target
mouseenter on container
mouseenter on host
mouseenter on hostParent

==Leaving target==
mouseleave on target
mouseleave on container
mouseleave on host
mouseleave on hostParent
